+++ src/SimplePie.php
@@ -1938,29 +1938,26 @@ class SimplePie
         }
 
         if (!$this->force_feed) {
+            // Check if the supplied URL is a feed, if it isn't, look for it.
+            $locate = $this->registry->create(Locator::class, [
+                &$file,
+                $this->timeout,
+                $this->useragent,
+                $this->max_checked_feeds,
+                $this->force_fsockopen,
+                $this->curl_options,
+                $this->get_http_client()
+            ]);
+
             $http_client = $this->get_http_client();
 
             if ($http_client instanceof Psr18Client) {
-                $locator_arguments = [
-                    &$file,
+                $locate->set_http_client(
                     $http_client->getHttpClient(),
                     $http_client->getRequestFactory(),
                     $http_client->getUriFactory()
-                ];
-            } else {
-                // BC: Pass deprecated file data
-                $locator_arguments = [
-                    &$file,
-                    $this->timeout,
-                    $this->useragent,
-                    $this->max_checked_feeds,
-                    $this->force_fsockopen,
-                    $this->curl_options
-                ];
+                );
             }
-
-            // Check if the supplied URL is a feed, if it isn't, look for it.
-            $locate = $this->registry->create(Locator::class, $locator_arguments);
 
             if (!$locate->is_feed($file)) {
                 $copyStatusCode = $file->get_status_code();